# openapi/community.json
{
  "components": {
    "schemas": {
...
      "Content": {
        "additionalProperties": false,
        "description": "Contains the multi-part content of a message.",
        "properties": {
          "parts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Part"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "List of parts that constitute a single message. Each part may have\n      a different IANA MIME type.",
            "title": "Parts"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. The producer of the content. Must be either 'user' or 'model'. If not set, the service will default to 'user'.",
            "title": "Role"
          }
        },
        "title": "Content",
        "type": "object"
      },
...
      "GenerateContentRequest": {
        "additionalProperties": false,
        "properties": {
          "cachedContent": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cachedcontent"
          },
          "contents": {
            "items": {
              "$ref": "#/components/schemas/Content"
            },
            "maxItems": 100000,
            "minItems": 1,
            "title": "Contents",
            "type": "array"
          },
          "generationConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "safetySettings": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SafetySetting"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safetysettings"
          },
          "serviceTier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceTier"
              },
              {
                "type": "null"
              }
            ]
          },
          "store": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store"
          },
          "systemInstruction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Content"
              },
              {
                "type": "null"
              }
            ]
          },
          "toolConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Tool"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tools"
          }
        },
        "required": [
          "contents"
        ],
        "title": "GenerateContentRequest",
        "type": "object"
      },
      "GenerationConfig": {
        "additionalProperties": false,
        "description": "Generation config.",
        "properties": {
          "audioTimestamp": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. If enabled, audio timestamps will be included in the request to the model. This can be useful for synchronizing audio with other modalities in the response. This field is not supported in Gemini API.",
            "title": "Audiotimestamp"
          },
          "audioTranscriptionConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AudioTranscriptionConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Configuration for audio transcription (speech recognition).\n      "
          },
          "candidateCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. The number of candidate responses to generate. A higher `candidate_count` can provide more options to choose from, but it also consumes more resources. This can be useful for generating a variety of responses and selecting the best one.",
            "title": "Candidatecount"
          },
          "enableAffectiveDialog": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. If enabled, the model will detect emotions and adapt its responses accordingly. For example, if the model detects that the user is frustrated, it may provide a more empathetic response. This field is not supported in Gemini API.",
            "title": "Enableaffectivedialog"
          },
          "enableEnhancedCivicAnswers": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Enables enhanced civic answers. It may not be available for all models. This field is not supported in Vertex AI.",
            "title": "Enableenhancedcivicanswers"
          },
          "frequencyPenalty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Penalizes tokens based on their frequency in the generated text. A positive value helps to reduce the repetition of words and phrases. Valid values can range from [-2.0, 2.0].",
            "title": "Frequencypenalty"
          },
          "logprobs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. The number of top log probabilities to return for each token. This can be used to see which other tokens were considered likely candidates for a given position. A higher value will return more options, but it will also increase the size of the response.",
            "title": "Logprobs"
          },
          "maxOutputTokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. The maximum number of tokens to generate in the response. A token is approximately four characters. The default value varies by model. This parameter can be used to control the length of the generated text and prevent overly long responses.",
            "title": "Maxoutputtokens"
          },
          "mediaResolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaResolution"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. The token resolution at which input media content is sampled. This is used to control the trade-off between the quality of the response and the number of tokens used to represent the media. A higher resolution allows the model to perceive more detail, which can lead to a more nuanced response, but it will also use more tokens. This does not affect the image dimensions sent to the model."
          },
          "modelSelectionConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelSelectionConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Config for model selection."
          },
          "presencePenalty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Penalizes tokens that have already appeared in the generated text. A positive value encourages the model to generate more diverse and less repetitive text. Valid values can range from [-2.0, 2.0].",
            "title": "Presencepenalty"
          },
          "responseFormat": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ResponseFormat"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. New response format field for the model to configure output formatting and delivery.",
            "title": "Responseformat"
          },
          "responseJsonSchema": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "description": "Output schema of the generated response. This is an alternative to\n      `response_schema` that accepts [JSON Schema](https://json-schema.org/).\n      ",
            "title": "Responsejsonschema"
          },
          "responseLogprobs": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. If set to true, the log probabilities of the output tokens are returned. Log probabilities are the logarithm of the probability of a token appearing in the output. A higher log probability means the token is more likely to be generated. This can be useful for analyzing the model's confidence in its own output and for debugging.",
            "title": "Responselogprobs"
          },
          "responseMimeType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. The IANA standard MIME type of the response. The model will generate output that conforms to this MIME type. Supported values include 'text/plain' (default) and 'application/json'. The model needs to be prompted to output the appropriate response type, otherwise the behavior is undefined. Deprecated: Use `response_format` instead.",
            "title": "Responsemimetype"
          },
          "responseModalities": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Modality"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. The modalities of the response. The model will generate a response that includes all the specified modalities. For example, if this is set to `[TEXT, IMAGE]`, the response will include both text and an image.",
            "title": "Responsemodalities"
          },
          "responseSchema": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Schema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Lets you to specify a schema for the model's response, ensuring that the output conforms to a particular structure. This is useful for generating structured data such as JSON. The schema is a subset of the [OpenAPI 3.0 schema object](https://spec.openapis.org/oas/v3.0.3#schema) object. When this field is set, you must also set the `response_mime_type` to `application/json`. Deprecated: Use `response_format` instead."
          },
          "routingConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationConfigRoutingConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Routing configuration. This field is not supported in Gemini API."
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. A seed for the random number generator. By setting a seed, you can make the model's output mostly deterministic. For a given prompt and parameters (like temperature, top_p, etc.), the model will produce the same response every time. However, it's not a guaranteed absolute deterministic behavior. This is different from parameters like `temperature`, which control the *level* of randomness. `seed` ensures that the \"random\" choices the model makes are the same on every run, making it essential for testing and ensuring reproducible results.",
            "title": "Seed"
          },
          "speechConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpeechConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. The speech generation config."
          },
          "stopSequences": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. A list of character sequences that will stop the model from generating further tokens. If a stop sequence is generated, the output will end at that point. This is useful for controlling the length and structure of the output. For example, you can use [\"\n\", \"###\"] to stop generation at a new line or a specific marker.",
            "title": "Stopsequences"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Controls the randomness of the output. A higher temperature results in more creative and diverse responses, while a lower temperature makes the output more predictable and focused. The valid range is (0.0, 2.0].",
            "title": "Temperature"
          },
          "thinkingConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThinkingConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Configuration for thinking features. An error will be returned if this field is set for models that don't support thinking."
          },
          "topK": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Specifies the top-k sampling threshold. The model considers only the top k most probable tokens for the next token. This can be useful for generating more coherent and less random text. For example, a `top_k` of 40 means the model will choose the next word from the 40 most likely words.",
            "title": "Topk"
          },
          "topP": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Specifies the nucleus sampling threshold. The model considers only the smallest set of tokens whose cumulative probability is at least `top_p`. This helps generate more diverse and less repetitive responses. For example, a `top_p` of 0.9 means the model considers tokens until the cumulative probability of the tokens to select from reaches 0.9. It's recommended to adjust either temperature or `top_p`, but not both.",
            "title": "Topp"
          },
          "translationConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranslationConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Config for translation. This field is not supported in Vertex AI."
          }
        },
        "title": "GenerationConfig",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MediaResolution": {
        "description": "The media resolution to use.",
        "enum": [
          "MEDIA_RESOLUTION_UNSPECIFIED",
          "MEDIA_RESOLUTION_LOW",
          "MEDIA_RESOLUTION_MEDIUM",
          "MEDIA_RESOLUTION_HIGH"
        ],
        "title": "MediaResolution",
        "type": "string"
      },
...
      "Modality": {
        "description": "Server content modalities.",
        "enum": [
          "MODALITY_UNSPECIFIED",
          "TEXT",
          "IMAGE",
          "AUDIO",
          "VIDEO"
        ],
        "title": "Modality",
        "type": "string"
      },
...
      "SafetySetting": {
        "additionalProperties": false,
        "description": "A safety setting that affects the safety-blocking behavior.\n\nA SafetySetting consists of a harm category and a threshold for that category.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HarmCategory"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required. The harm category to be blocked."
          },
          "method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HarmBlockMethod"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. The method for blocking content. If not specified, the default behavior is to use the probability score. This field is not supported in Gemini API."
          },
          "threshold": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HarmBlockThreshold"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required. The threshold for blocking content. If the harm probability exceeds this threshold, the content will be blocked."
          }
        },
        "title": "SafetySetting",
        "type": "object"
      },
...
      "ServiceTier": {
        "description": "Pricing and performance service tier.",
        "enum": [
          "unspecified",
          "flex",
          "standard",
          "priority"
        ],
        "title": "ServiceTier",
        "type": "string"
      },
...
      "Tool": {
        "additionalProperties": false,
        "description": "Tool details of a tool that the model may use to generate a response.",
        "properties": {
          "codeExecution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolCodeExecution"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. CodeExecution tool type. Enables the model to execute code as part of generation."
          },
          "computerUse": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComputerUse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Tool to support the model interacting directly with the computer. If enabled, it automatically populates computer-use specific Function Declarations."
          },
          "enterpriseWebSearch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EnterpriseWebSearch"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Tool to support searching public web data, powered by Vertex AI Search and Sec4 compliance. This field is not supported in Gemini API."
          },
          "exaAiSearch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolExaAiSearch"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Uses Exa.ai to search for information to answer user queries. The search results will be grounded on Exa.ai and presented to the model for response generation. This field is not supported in Gemini API."
          },
          "fileSearch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileSearch"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. FileSearch tool type. Tool to retrieve knowledge from Semantic Retrieval corpora. This field is not supported in Vertex AI."
          },
          "functionDeclarations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FunctionDeclaration"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Function tool type. One or more function declarations to be passed to the model along with the current user query. Model may decide to call a subset of these functions by populating FunctionCall in the response. User should provide a FunctionResponse for each function call in the next turn. Based on the function responses, Model will generate the final response back to the user. Maximum 512 function declarations can be provided.",
            "title": "Functiondeclarations"
          },
          "googleMaps": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoogleMaps"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Tool that allows grounding the model's response with\n      geospatial context related to the user's query."
          },
          "googleSearch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoogleSearch"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. GoogleSearch tool type. Tool to support Google Search in Model. Powered by Google."
          },
          "googleSearchRetrieval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoogleSearchRetrieval"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Specialized retrieval tool that is powered by Google Search."
          },
          "mcpServers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/McpServer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. MCP Servers to connect to. This field is not supported in Vertex AI.",
            "title": "Mcpservers"
          },
          "parallelAiSearch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolParallelAiSearch"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. If specified, Vertex AI will use Parallel.ai to search for information to answer user queries. The search results will be grounded on Parallel.ai and presented to the model for response generation. This field is not supported in Gemini API."
          },
          "retrieval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Retrieval"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Retrieval tool type. System will always execute the provided retrieval tool(s) to get external knowledge to answer the prompt. Retrieval results are presented to the model for generation. This field is not supported in Gemini API."
          },
          "urlContext": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UrlContext"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Tool to support URL context retrieval."
          }
        },
        "title": "Tool",
        "type": "object"
      },
...
      "ToolConfig": {
        "additionalProperties": false,
        "description": "Tool config.\n\nThis config is shared for all tools provided in the request.",
        "properties": {
          "functionCallingConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FunctionCallingConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Function calling config."
          },
          "includeServerSideToolInvocations": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. If true, the API response will include the server-side tool calls and responses within the `Content` message. This allows clients to observe the server's tool interactions. This field is not supported in Vertex AI.",
            "title": "Includeserversidetoolinvocations"
          },
          "retrievalConfig": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Retrieval config."
          }
        },
        "title": "ToolConfig",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AnthropicAPIKey": {
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "ShimAPIKey": {
        "in": "header",
        "name": "x-shim-key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "SHIM Trust Boundary Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1beta/models/{model}:generateContent": {
      "post": {
        "operationId": "generate_content_v1beta_models__model__generateContent_post",
        "parameters": [
          {
            "in": "path",
            "name": "model",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
              "title": "Model",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ShimAPIKey": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Content"
      }
    },
    "/v1beta/models/{model}:streamGenerateContent": {
      "post": {
        "operationId": "stream_generate_content_v1beta_models__model__streamGenerateContent_post",
        "parameters": [
          {
            "in": "path",
            "name": "model",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
              "title": "Model",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "alt",
            "required": true,
            "schema": {
              "const": "sse",
              "title": "Alt",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Native Gemini server-sent events."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ShimAPIKey": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stream Generate Content"
      }
    }
  }
}